Generalise updater.go and move into its own package
